Add ard_submission.ipynb via upload
This notebook can successfully take a csv obtained by the ARD through eSirius and populate an instance of MariaDB hosted locally on Jeremy Delahanty's laptop. It can parse the csv for orders relevant to the project's investigator and then populate the database with automatically generated cage ids containing the appropriate number of mice defined in the order.